+++ src/callable/primitive/paste.rs
@@ -2,6 +2,7 @@ use r_derive::*;
 
 use crate::error::*;
 use crate::lang::*;
+use crate::context::Context;
 use crate::callable::core::*;
 use crate::object::*;
