+++ src/lang.rs
@@ -1,4 +1,5 @@
 use crate::callable::core::{builtin, Callable};
+use crate::context::Context;
 use crate::error::*;
 use crate::internal_err;
 use crate::object::types::*;
@@ -69,7 +70,7 @@ impl Obj {
             // don't evaluate in a new frame, but rather just in originating
             // environment?
             Obj::Closure(expr, env) => {
-                stack.add_frame(expr.clone(), env);
+                stack.add_frame(expr.clone(), env.clone());
                 let result = stack.eval(expr);
                 stack.pop_frame_and_return(result)
             }
@@ -233,6 +234,7 @@ impl Obj {
         }
     }
 
+    // Used for [ ] syntax
     pub fn try_get(&self, index: Obj) -> EvalResult {
         match self {
             Obj::Vector(v) => v.try_get(index),
@@ -241,6 +243,7 @@ impl Obj {
         }
     }
 
+    // Used for [[ ]] syntax
     pub fn try_get_inner(&self, index: Obj) -> EvalResult {
         match self {
             Obj::Vector(v) => v.try_get(index),
@@ -536,11 +539,11 @@ impl Frame {
         Self { call, to, env }
     }
 
-    pub fn new_child_env(&self) -> Rc<Environment> {
-        Rc::new(Environment {
-            parent: Some(self.env.clone()),
+    pub fn new_child_env(&self) -> Box<dyn Context> {
+        Box::new(Obj::Environment(Rc::new(Environment {
+            parent: Some(self.env().clone()),
             ..Default::default()
-        })
+        })))
     }
 }
 
@@ -577,17 +580,17 @@ impl CallStack {
         }
     }
 
-    pub fn last_frame(&self) -> &Frame {
+    pub fn last_frame(&self) -> Frame {
         if let Some(frame) = self.frames.last() {
-            frame
+            frame.clone()
         } else {
             panic!("We've somehow exhausted the entire call stack and are still evaluating")
         }
     }
 
-    pub fn parent_frame(&self) -> &Frame {
+    pub fn parent_frame(&self) -> Frame {
         if let Some(frame) = self.frame(-1) {
-            frame
+            frame.clone()
         } else {
             panic!("Attempting access to parent frame at top level")
         }
@@ -611,7 +614,7 @@ impl CallStack {
 impl Display for CallStack {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         for (i, frame) in self.frames.iter().enumerate().skip(1) {
-            writeln!(f, "{}: {} {}", i, frame.call, frame.env.clone())?;
+            writeln!(f, "{}: {} {}", i, frame.call, frame.clone())?;
         }
         Ok(())
     }
@@ -630,93 +633,6 @@ impl From<Rc<Environment>> for CallStack {
         CallStack {
             frames: vec![Frame::new(Expr::Null, value.clone())],
         }
-    }
-}
-
-pub trait Context {
-    fn get(&mut self, name: String) -> EvalResult {
-        (*self).env().get(name)
-    }
-
-    fn get_ellipsis(&mut self) -> EvalResult {
-        let err = Err(Signal::Error(RError::IncorrectContext("...".to_string())));
-        self.get("...".to_string()).or(err)
-    }
-
-    fn assign_lazy(&mut self, _to: Expr, _from: Expr) -> EvalResult {
-        Err(Signal::Error(RError::IncorrectContext("<-".to_string())))
-    }
-
-    fn assign(&mut self, _to: Expr, _from: Obj) -> EvalResult {
-        Err(Signal::Error(RError::IncorrectContext("<-".to_string())))
-    }
-
-    fn env(&self) -> Rc<Environment>;
-
-    fn eval_call(&mut self, expr: Expr) -> EvalResult {
-        self.eval(expr)
-    }
-
-    fn eval(&mut self, expr: Expr) -> EvalResult {
-        self.env().eval(expr)
-    }
-
-    fn eval_and_finalize(&mut self, expr: Expr) -> EvalResult {
-        self.eval(expr)
-    }
-
-    fn eval_binary(&mut self, exprs: (Expr, Expr)) -> Result<(Obj, Obj), Signal> {
-        Ok((self.eval(exprs.0)?, self.eval(exprs.1)?))
-    }
-
-    fn eval_list_lazy(&mut self, l: ExprList) -> EvalResult {
-        Ok(Obj::List(List::from(
-            l.into_iter()
-                .flat_map(|pair| match pair {
-                    (_, Expr::Ellipsis(None)) => {
-                        if let Ok(Obj::List(ellipsis)) = self.get_ellipsis() {
-                            ellipsis.values.borrow_mut().clone().into_iter()
-                        } else {
-                            vec![].into_iter()
-                        }
-                    }
-                    (k, e @ (Expr::Call(..) | Expr::Symbol(..))) => {
-                        let elem = vec![(k, Obj::Closure(e, self.env()))];
-                        elem.into_iter()
-                    }
-                    (k, v) => {
-                        if let Ok(elem) = self.eval(v) {
-                            vec![(k, elem)].into_iter()
-                        } else {
-                            unreachable!()
-                        }
-                    }
-                })
-                .collect::<Vec<_>>(),
-        )))
-    }
-
-    fn eval_list_eager(&mut self, l: ExprList) -> EvalResult {
-        Ok(Obj::List(List::from(
-            l.into_iter()
-                .map(|pair| match pair {
-                    (_, Expr::Ellipsis(None)) => {
-                        if let Ok(Obj::List(ellipsis)) = self.get_ellipsis() {
-                            Ok(ellipsis.values.borrow_mut().clone().into_iter())
-                        } else {
-                            Ok(vec![].into_iter())
-                        }
-                    }
-                    (k, v) => match self.eval(v) {
-                        Ok(elem) => Ok(vec![(k, elem)].into_iter()),
-                        Err(e) => Err(e),
-                    }
-                })
-                .collect::<Result<Vec<_>, _>>()?
-                .into_iter()
-                .flatten()
-                .collect::<Vec<_>>(),
-        )))
     }
 }
 
@@ -782,7 +698,7 @@ impl Context for CallStack {
     }
 
     fn env(&self) -> Rc<Environment> {
-        self.last_frame().env.clone()
+        self.last_frame().env().clone()
     }
 
     fn eval_call(&mut self, expr: Expr) -> EvalResult {
@@ -919,9 +835,45 @@ impl Context for CallStack {
     // Try.
 }
 
-impl Context for &Frame {
+impl Context for Frame {
     fn env(&self) -> Rc<Environment> {
         self.env.clone()
+    }
+}
+
+impl Context for Obj {
+    fn env(&self) -> Rc<Environment> {
+        match self {
+            Obj::Environment(e) => e.clone(),
+            _ => unimplemented!(),
+        }
+    }
+
+    fn eval(&mut self, expr: Expr) -> EvalResult {
+        match expr {
+            Expr::Null => Ok(Obj::Null),
+            Expr::NA => Ok(Obj::Vector(Vector::from(vec![OptionNA::NA as Logical]))),
+            Expr::Inf => Ok(Obj::Vector(Vector::from(vec![OptionNA::Some(f64::INFINITY)]))),
+            Expr::Number(x) => Ok(Obj::Vector(Vector::from(vec![x]))),
+            Expr::Integer(x) => Ok(Obj::Vector(Vector::from(vec![x]))),
+            Expr::Bool(x) => Ok(Obj::Vector(Vector::from(vec![OptionNA::Some(x)]))),
+            Expr::String(x) => Ok(Obj::Vector(Vector::from(vec![OptionNA::Some(x)]))),
+            Expr::Function(formals, body) => Ok(Obj::Function(formals, *body, self.env().clone())),
+            Expr::Symbol(name) => self.get(name),
+            Expr::Break => Err(Signal::Condition(Cond::Break)),
+            Expr::Continue => Err(Signal::Condition(Cond::Continue)),
+            Expr::Primitive(p) => Ok(Obj::Function(p.formals(), Expr::Primitive(p), self.environment().unwrap())),
+            Expr::More => Ok(Obj::Null),
+            x => internal_err!(format!("Can't evaluate Context::eval(Rc<Envrionment>, {x}")),
+        }
+    }
+
+    fn get(&mut self, name: String) -> EvalResult {
+        match self {
+            Obj::List(l) => l.try_get_inner(Obj::Vector(Vector::from(vec![name]))),
+            Obj::Environment(e) => e.get(name),
+            _ => unimplemented!(),
+        }
     }
 }
 
@@ -934,14 +886,12 @@ impl Context for Rc<Environment> {
         match expr {
             Expr::Null => Ok(Obj::Null),
             Expr::NA => Ok(Obj::Vector(Vector::from(vec![OptionNA::NA as Logical]))),
-            Expr::Inf => Ok(Obj::Vector(Vector::from(vec![OptionNA::Some(
-                f64::INFINITY,
-            )]))),
+            Expr::Inf => Ok(Obj::Vector(Vector::from(vec![OptionNA::Some(f64::INFINITY)]))),
             Expr::Number(x) => Ok(Obj::Vector(Vector::from(vec![x]))),
             Expr::Integer(x) => Ok(Obj::Vector(Vector::from(vec![x]))),
             Expr::Bool(x) => Ok(Obj::Vector(Vector::from(vec![OptionNA::Some(x)]))),
             Expr::String(x) => Ok(Obj::Vector(Vector::from(vec![OptionNA::Some(x)]))),
-            Expr::Function(formals, body) => Ok(Obj::Function(formals, *body, self.clone())),
+            Expr::Function(formals, body) => Ok(Obj::Function(formals, *body, self.env().clone())),
             Expr::Symbol(name) => self.get(name),
             Expr::Break => Err(Signal::Condition(Cond::Break)),
             Expr::Continue => Err(Signal::Condition(Cond::Continue)),
@@ -952,29 +902,6 @@ impl Context for Rc<Environment> {
     }
 
     fn get(&mut self, name: String) -> EvalResult {
-        // search in this environment for value by name
-        if let Some(value) = self.values.borrow().get(&name) {
-            let result = value.clone();
-            return match result {
-                Obj::Closure(expr, mut env) => env.eval(expr),
-                _ => Ok(result),
-            };
-
-        // if not found, search through parent if available
-        } else if let Some(parent) = &self.parent {
-            parent.clone().get(name)
-
-        // if we're at the top level, fall back to primitives if available
-        } else if let Ok(prim) = name.as_str().try_into() {
-            Ok(Obj::Function(
-                ExprList::new(),
-                Expr::Primitive(prim),
-                self.env(),
-            ))
-
-        // otherwise, throw error
-        } else {
-            Err(Signal::Error(RError::VariableNotFound(name)))
-        }
+        Environment::get(self, name)
     }
 }
