+++ src/lib.rs
@@ -4,9 +4,11 @@ extern crate pest_derive;
 pub mod utils;
 
 pub mod error;
-pub mod lang;
 pub mod parser;
 
+pub mod context;
+pub mod lang;
 pub mod callable;
+
 pub mod repl;
 pub mod object;
